+++ src/DependencyInjection/SonataUserExtension.php
@@ -59,7 +59,7 @@ class SonataUserExtension extends Extension implements PrependExtensionInterface
 
         $loader->load('twig.xml');
         $loader->load('actions.xml');
-        $loader->load('event_listener.xml');
+        $loader->load('listener.xml');
         $loader->load('mailer.xml');
         $loader->load('form.xml');
         $loader->load('security.xml');
